// src/checkout/components/address-form/types.ts
import { type AddressFragment, type CountryCode } from "@/checkout/graphql";
import { type MightNotExist } from "@/checkout/lib/global-types";

export interface AddressFormData extends Omit<Record<AddressField, string>, "country" | "countryCode"> {
	countryCode: CountryCode;
}

export type OptionalAddress = MightNotExist<AddressFragment>;

export type AddressField =
	| "city"
	| "firstName"
	| "lastName"
	| "countryArea"
	| "cityArea"
	| "postalCode"
	| "countryCode"
	| "companyName"
	| "streetAddress1"
	| "streetAddress2"
	| "phone";

export type ApiAddressField = AddressField | "name";